Make headers show up in IDE project
In CMakeLists.txt, create a custom target using add_custom_target() and
list the headers of header-only modules as the sources of the custom
target in order to have the headers show up in the IDE projects
generated by CMake. This only needs to be done for headers of
header-only modules. Headers that have a corresponding .cpp file will
be included automatically by CMake.